requirements.txt
piecash[finance-quote,postgres]==0.19.0
alpha_vantage==2.1.0
python-dateutil